fx quantization: split patterns into separate file
Summary:

The `quantize.py` file is hard to read because of high LOC.
Splitting the patterns into their own file so files are
more single purposed.

Test Plan:

```
python test/test_quantization.py TestQuantizeFx
python test/test_quantization.py TestQuantizeFxOps
```

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
